blog_app_backend/requirements.txt
Django==4.2.4
djangorestframework==3.15.0
psycopg2
psycopg2-binary
python-dotenv
django-rest-authtoken
faker
django-cors-headers
